gui: save memory in Segment

+++ src/grt/src/fastroute/include/FastRoute.h
@@ -108,7 +108,7 @@ class FastRouteCore
                 int min_layer,
                 int max_layer,
                 float slack,
-                std::vector<int>* edge_cost_per_layer);
+                std::vector<int8_t>* edge_cost_per_layer);
   void deleteNet(odb::dbNet* db_net);
   void removeNet(odb::dbNet* db_net);
   void mergeNet(odb::dbNet* removed_net, odb::dbNet* preserved_net);
